Put custom fluids in the core namespace and give them names
