+++ common/src/main/resources/assets/logistics/lang/en_us.json
@@ -9,6 +9,11 @@
   "block.logistics.pipe.copper_transport_pipe.waxed.exposed": "Waxed Exposed Copper Transport Pipe",
   "block.logistics.pipe.copper_transport_pipe.waxed.weathered": "Waxed Weathered Copper Transport Pipe",
   "block.logistics.pipe.copper_transport_pipe.waxed.oxidized": "Waxed Oxidized Copper Transport Pipe",
+  "fluid.logistics.core.liquid_redstone": "Liquid Redstone",
+  "fluid.logistics.core.liquid_ender": "Liquid Ender",
+  "fluid.logistics.core.liquid_glowstone": "Liquid Glowstone",
+  "fluid.logistics.core.crude_oil": "Crude Oil",
+  "fluid.logistics.core.liquid_biomass": "Liquid Biomass",
   "block.logistics.fluid.copper_fluid_pipe": "Copper Fluid Pipe",
   "block.logistics.fluid.copper_fluid_pipe.exposed": "Exposed Copper Fluid Pipe",
   "block.logistics.fluid.copper_fluid_pipe.weathered": "Weathered Copper Fluid Pipe",
